Extract text manipulation helpers
